chore: Update `DatabaseProductDownloadFromBiketadeJob` to use `Batchable` trait
The `DatabaseProductDownloadFromBiketadeJob` class now uses the `Batchable` trait in addition to other traits (`Dispatchable`, `InteractsWithQueue`, `Queueable`, `SerializesModels`). This change allows the job to be included in a batch of jobs for improved performance and management.

Note: This change is related to the recent addition of the job for downloading products from Biketade.