chore: Update `DatabaseProductDownloadFromBikeFunJob` to download and update product data from CSV
This commit updates the `DatabaseProductDownloadFromBikeFunJob` class to download the product data from a CSV file and update the corresponding records in the database. The CSV file is fetched from the URL `https://xml.bikefun.hu/cikktorzs.csv`. The downloaded data is parsed and processed to update the `Product` model attributes such as price, stock, product name, URL picture, barcode, and description. This change improves the functionality of the application by keeping the product data up to date.